Refactored NEVP tests
Such that it uses the same data as the obc tests
Additionally, removed testing for the left eigenvectors
